Moar testz. Moved traffic light + stu machine into separate files for clarify
