+++ tests/protocol/statemachine/integration_state.py
@@ -1,0 +1,8 @@
+from unittest import TestCase
+from unittest.mock import MagicMock
+from .trafficlight import *
+from cilantro.utils.test import MPTestCase, mp_testable
+
+
+class IntegrationTestState(MPTestCase):
+    pass
